Initial commit
The program is basically a prototype at this point, but the design is
good enough to iterate on.

The front-end (www/client/) needs a lot of work.  It isn't even possible
to navigate it without typing in URLs directly, and the pages are
completely un-styled.

The back-end has the basic features complete, but there's probably lots
of bugs lurking there.  There are still some very necessary features to
implement, such as authentication.